下单操作 购物车实现

+++ mall_service_api/mall_service_order_api/src/main/java/com/ytz/mall/order/pojo/ReturnOrderItem.java
@@ -91,7 +91,7 @@ public class ReturnOrderItem implements Serializable{
 	 * 重量
 	 */
     @Column(name = "weight")
-	private Integer weight;
+	private BigDecimal weight;
 
 
 }
